feat(test): 重整stringx目录测试文件结构
将8个测试文件重组为5个标准命名的测试文件，遵循源文件对应规范：
- 删除孤立文件: benchmark_test.go, complete_coverage_test.go, missing_coverage_test.go
- 增强文件: rand_test.go, string_test.go, unicode_test.go, utf16_test.go
- 合并基准测试和覆盖率测试到相应源文件的测试中
- 保持100%测试功能并增强性能基准测试
